Clean up `RazorFilePathToContentTypeProviders` a bit
